+++ memory_update_protocol.py
